add more tests for AppMode & AppState stringConvertible
